feedback

+++ packages/protocol/contracts/L1/TaikoData.sol
@@ -125,7 +125,9 @@ library TaikoData {
         // this block is not verified as the last block in a batch, verifiedTransitionId
         // will remain zero.
         uint32 verifiedTransitionId;
-        uint64 l1StateBlockNumber;
+        // This block number is the value for _l1BlockId (and related values) in TaikoL2's anchor()
+        // function.
+        uint32 l1StateBlockNumber; // slot 4
     }
 
     /// @dev Struct representing an Ethereum deposit.
